split Process.ts
